Cut out nanopub functionalities, they are moved to a separate library
