+++ MANIFEST.in
@@ -1,6 +1,5 @@
 include README.md
 include CITATION.cff
-include fairworkflows/np
 include fairworkflows/_version.py
 include requirements.txt
 include requirements_dev.txt
