+++ fairworkflows/__init__.py
@@ -1,5 +1,3 @@
 from ._version import __version__
 from .fairstep import FairStep
-from .fairworkflow import FairWorkflow
-from .decorators import fairstep
-from .nanopub import Nanopub
+from .fairworkflow import FairWorkflow, add_step
