--- tests/resources/nanopub_sample_unsigned.trig
@@ -1,56 +1,0 @@
-@prefix this: <'http://purl.org/nanopub/temp/mynanopub#'> .
-@prefix sub: <http://purl.org/np/RAzPytdERsBd378zHGvwgRbat1MCiS7QrxNrPxe9yDu6E#> .
-@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-@prefix dct: <http://purl.org/dc/terms/> .
-@prefix dce: <http://purl.org/dc/elements/1.1/> .
-@prefix pav: <http://purl.org/pav/> .
-@prefix np: <http://www.nanopub.org/nschema#> .
-@prefix npx: <http://purl.org/nanopub/x/> .
-
-sub:Head {
-  this: np:hasAssertion sub:assertion;
-    np:hasProvenance sub:provenance;
-    np:hasPublicationInfo sub:pubinfo;
-    a np:Nanopublication .
-}
-
-sub:assertion {
-  this: npx:includesElement <http://purl.org/np/RA1o-WSjHkm56DZuzziAYoUmuF-muVtPdlQKvsCT4yaPA>,
-      <http://purl.org/np/RA1q1I6-EtcuWMVsCZECCASXoiWt61VMOgDJ9JSCP4UGU>, <http://purl.org/np/RA4m_1AFmib-aDjLwgKLfIuxKEqa0pHh5H4cyc8_czWx0>,
-      <http://purl.org/np/RA58xJdnse4hTrX8IdNqgPYLTW_kHdIeB14qiTtmwFvCo>, <http://purl.org/np/RA7HMrJxgphlwIVFoQOjKGGSAYrFIxZyoau9lTJvub38A>,
-      <http://purl.org/np/RA9dUlTqvAT0LumWwa6UX_Cztck2J6w55L91HdvMjjppc>, <http://purl.org/np/RA9nG0iq5jiyVIQz30GbxLWbxnFEUFqLe5_IA5yJEW7Qo>,
-      <http://purl.org/np/RAAhmGwEay7buLOKE4OQ1erwlBK9TYuAcJq1eW6yfBOO8>, <http://purl.org/np/RAB8B41GzXHUN3VzlWFwQRh_q7WSugLI7jZkNFWAe1Kxs>,
-      <http://purl.org/np/RABcjCt6abFWWO24CnUjYXE9OvphOOdjHP-pTgV7dmaKI>, <http://purl.org/np/RABhqk7zczoyyvYgzc-PokSkWL59f9ssXvr3No000cc-Q>,
-      <http://purl.org/np/RABkQsWsWGNXay34cUMFD3K86nQNkCBuCPTfdxYxtWdUg>, <http://purl.org/np/RACOKd9QOjLI4VMY4N7f9iMYgFrZk7vAMJumXRol6wmrQ>,
-      <http://purl.org/np/RAE-AEj4KkkO8-cNUVs_WJbLhUuuYEWhyvQq3vseC2gbU>, <http://purl.org/np/RAEbrpJ2n7g6TXCsP9uxEFb090obEDjDjWgyKH_HteFO0>,
-      <http://purl.org/np/RAFVQa7-vWMpmyZGzM6IOGhjHjNhv6mlNQE5Hpf6n-4Q8>, <http://purl.org/np/RAFd7nBvLeyisKDKhDEymlIVg7GgjubN4O8wx5gbtZ7oE>,
-      <http://purl.org/np/RAKkRx76v2pge1qSti0NPFOEHNhVYkaLsRKKPHS6pd-oE>, <http://purl.org/np/RAMjlvyD7CYRArrnOYNpFANje9LtXNwzMpmAzaWihj4bM>,
-      <http://purl.org/np/RAO3HmwRoDFMCJ0y-_KgYh1mbFTlYMlNG1GO3d6QyPXLg>, <http://purl.org/np/RAP-vr4oeD-JO7d_XrTG2QtPxaOIPVrQasG-GviCBl97A>,
-      <http://purl.org/np/RAP8AHpEim15EIxOfZnuG9xpSmERPH7eHfvj52G0197lY>, <http://purl.org/np/RAP8eTaGW5nZzpabQ7yM7G9l3ynl_nNLfKXn3uGlzl80I>,
-      <http://purl.org/np/RAPkivO794VVcqpBI6zvfddkKyVJcQbHNBCrkr3C8jxVQ>, <http://purl.org/np/RAQ8zOgtjLzzyjsA41EF4Zj_4TbUStmHYJeo7APx6WXtE>,
-      <http://purl.org/np/RARvRtlCJK2pZZA285npQBSRsXIuW0GMOMl2CpzvIODOM>, <http://purl.org/np/RASNtMTdARdcYtB3TTuPClrmedeDJOMnsAesS7UokiChY>,
-      <http://purl.org/np/RATaBeXjiybRjZajYnb7KOqKtCfTpWtSSq6FcUDaDTTDw>, <http://purl.org/np/RATvnu4oFZfgRdf7YwEhCH5ZmNhrN2b-DLjk9UyZndnhc>,
-      <http://purl.org/np/RAVfXk2uPdKBEgfDEKqqBUY-tcQ5nlJi4gvTJNKBwa6hQ>, <http://purl.org/np/RAYpwK2628KMvA6Kzrk3KG09IW8KkwFh_HcHDDijc2En0>,
-      <http://purl.org/np/RA_K6UmHfzBwoA4VFHiuitdzeCm-huphxH-RSmuZTIdZw>, <http://purl.org/np/RAa0---1HK9tDOZklpPWakJ50uHAqwPw9bKkmC2UdAf8A>,
-      <http://purl.org/np/RAbOlaf0fKUSlhYH3IT9_HBGhGRuYcMKWdn0EcYyftyrs>, <http://purl.org/np/RAcMpiliHJP4htUmEdI9zCTs22smZfjPNNTnReBtD9Hlw>,
-      <http://purl.org/np/RAcTAtn3uLHzpl9A6lWQkUk6jwH53-cBlDXzL70f53eus>, <http://purl.org/np/RAcyIgMk85Be-1wrAAR4dvfLDP0WMv4Mz_PnDfH-33gsU>,
-      <http://purl.org/np/RAdZiZ3JJWqPV1J5PfLc8NBrE8zTgjARvj0uXxn0Dat-A>, <http://purl.org/np/RAgdJQTutzID92DI29H_trr-tx8OB1J5TD6cM4Fi-11mI>,
-      <http://purl.org/np/RAip5sNsf91mktx8dahv43TZvx8MnZ55eRr303pdEQUyc>, <http://purl.org/np/RAk6-MXW5fdn28jXWCS84e0e8TRmkGyGwq3CKGqyvs6hM>,
-      <http://purl.org/np/RAkf3rzncTOot7nOHAlhvYjW-UXZYo4GpwcUVIDOV_X34>, <http://purl.org/np/RAnc71A2WKBP43w_yOGdikUGoMZWsHOEzKK56se6dw4Eg>,
-      <http://purl.org/np/RApiM5Rj8Fb0VEbg2w5wI5lyU4ajhnGMJ39tWwRssLi2k>, <http://purl.org/np/RArnuHhoNY934aeD2N_wQRGDDirXkbdMBSk5eOMS--qPw>,
-      <http://purl.org/np/RAt-X373jf8txtzNGJben9qkOVoS4GCG66sUpTi35rnJY>, <http://purl.org/np/RAt3jp0MnXmUvo6UnlOupHFj4ZVj40UQITg-Qy5WD9_GQ>,
-      <http://purl.org/np/RAvld4We77jP84OAWsMqpEVGc1lDrJ0qV7VJCrJKINguk>, <http://purl.org/np/RAvpasBCfFWMWjnV47eryXvYOBrdyymtE1p9G6lOfYpPU> .
-}
-
-sub:provenance {
-  sub:assertion a npx:IndexAssertion .
-}
-
-sub:pubinfo {
-
-  this: dce:description " index contains a collection of nanopublications that represent a network of claims concerning physician suicide rate in the United States. The text of the claim is indicated along with any citations provided for that claim. In some cases, the cited articles do not make any kind of assertion of that nature, and a comment is used to indicate it.";
-    dce:title "US Physician Suicide Rate Claim Network";
-    dct:created "2020-04-30T10:01:24.756Z"^^xsd:dateTime;
-    pav:createdBy <https://orcid.org/0000-0002-6007-4023>, <https://orcid.org/0000-0003-4727-9435>;
-    a npx:NanopubIndex .
-}
